FAI-1092 point trainingCourse post and put endpoints to the same upsert put endpoint in candidates
