+++ src/FindAnApprenticeship/SFA.DAS.FindAnApprenticeship.Api.UnitTests/Controllers/TrainingCoursesController/WhenPostingTrainingCourse.cs
@@ -20,7 +20,6 @@ public class WhenPostingTrainingCourse
     public async Task Then_The_Command_Response_Is_Returned(
         Guid applicationId,
         PostTrainingCourseApiRequest apiRequest,
-        CreateTrainingCourseCommandResponse response,
         [Frozen] Mock<IMediator> mediator,
         [Greedy] Api.Controllers.TrainingCoursesController controller)
     {
@@ -30,17 +29,11 @@ public class WhenPostingTrainingCourse
             && c.CourseName == apiRequest.CourseName
             && c.YearAchieved == apiRequest.YearAchieved),
             CancellationToken.None))
-            .ReturnsAsync(response);
+            .ReturnsAsync(Unit.Value);
 
-        var actual = await controller.PostTrainingCourse(applicationId, apiRequest) as CreatedResult;
-        var actualModel = actual?.Value as PostTrainingCourseApiResponse;
+        var actual = await controller.PostTrainingCourse(applicationId, apiRequest);
 
-        using (new AssertionScope())
-        {
-            actual.Should().NotBeNull();
-            actualModel.Should().NotBeNull();
-            actualModel.Should().BeEquivalentTo((PostTrainingCourseApiResponse)response);
-        }
+        actual.Should().BeOfType<OkResult>();
     }
 
     [Test, MoqAutoData]
@@ -58,25 +51,6 @@ public class WhenPostingTrainingCourse
         {
             actual.Should().NotBeNull();
             actual.StatusCode.Should().Be((int)HttpStatusCode.InternalServerError);
-        }
-    }
-
-    [Test, MoqAutoData]
-    public async Task And_Null_Is_Returned_Then_Not_Found_Response_Returned(
-        Guid applicationId,
-        PostTrainingCourseApiRequest apiRequest,
-        [Frozen]
-        Mock<IMediator> mediator,
-        [Greedy] Api.Controllers.TrainingCoursesController controller)
-    {
-        mediator.Setup(x => x.Send(It.IsAny<CreateTrainingCourseCommand>(), CancellationToken.None)).ReturnsAsync(() => null);
-
-        var actual = await controller.PostTrainingCourse(applicationId, apiRequest) as StatusCodeResult;
-
-        using (new AssertionScope())
-        {
-            actual.Should().NotBeNull();
-            actual.StatusCode.Should().Be((int)HttpStatusCode.NotFound);
         }
     }
 }
