+++ src/FindAnApprenticeship/SFA.DAS.FindAnApprenticeship.Api/Models/Applications/PostTrainingCourseApiRequest.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using SFA.DAS.FindAnApprenticeship.Application.Commands.Apply.CreateTrainingCourse;
 
 namespace SFA.DAS.FindAnApprenticeship.Api.Models.Applications;
 
@@ -8,17 +7,4 @@ public class PostTrainingCourseApiRequest
     public Guid CandidateId { get; set; }
     public string CourseName { get; set; }
     public int YearAchieved { get; set; }
-}
-
-public class PostTrainingCourseApiResponse
-{
-    public Guid Id { get; set; }
-
-    public static implicit operator PostTrainingCourseApiResponse(CreateTrainingCourseCommandResponse source)
-    {
-        return new PostTrainingCourseApiResponse
-        {
-            Id = source.Id
-        };
-    }
 }
