+++ src/FindAnApprenticeship/SFA.DAS.FindAnApprenticeship.UnitTests/InnerApi/CandidateApi/Requests/WhenBuildingPutUpdateTrainingCourseApiRequest.cs
@@ -12,7 +12,7 @@ public class WhenBuildingPutUpdateTrainingCourseApiRequest
             Guid candidateId,
             Guid trainingCourseId)
     {
-        var actual = new PutUpdateTrainingCourseApiRequest(applicationId, candidateId, trainingCourseId, new PutUpdateTrainingCourseApiRequest.PutUpdateTrainingCourseApiRequestData());
+        var actual = new PutUpsertTrainingCourseApiRequest(applicationId, candidateId, trainingCourseId, new PutUpsertTrainingCourseApiRequest.PutUpdateTrainingCourseApiRequestData());
 
         actual.PutUrl.Should().Be($"candidates/{candidateId}/applications/{applicationId}/trainingcourses/{trainingCourseId}");
     }
