--- src/FindAnApprenticeship/SFA.DAS.FindAnApprenticeship/InnerApi/CandidateApi/Responses/PostTrainingCourseApiResponse.cs
@@ -1,7 +1,0 @@
-﻿using System;
-
-namespace SFA.DAS.FindAnApprenticeship.InnerApi.CandidateApi.Responses;
-public class PostTrainingCourseApiResponse
-{
-    public Guid Id { get; set; }
-}
